Write CNAME file if custom domain
Custom domains in GitHub pages require a CNAME file. This file is automatically created for you if you add a custom
domain to the GitHub pages section of the "Settings" area in a GitHub repo.

However, the way we push to GitHub, we remove all files from a `gh-pages` branch before building and pushing the new
content on a rebuild. This includes removing the CNAME file, even if it was autogenerated when changing the setting.

So, just always create a CNAME file if the domain is a custom domain.