Update getting started page to be more intuitive (#900)
* Updating step 2 to be more intuitive

Use of word Github is very confusing. When starting, most of devs wont have an existing github repo. And normally not everyone will be using Github. Since devs will be following this step by step: we need to stress on telling them to create a new or use existing local project folder which may or may not be a github repo.

* Update getting-started-installation.md